加入userInfo内容

+++ app/src/main/java/com/fr1014/mycoludmusic/data/entity/http/wangyiyun/WYUserPlayList.kt
@@ -50,7 +50,7 @@ data class Playlist(
     val tracks: Any,
     val updateFrequency: Any,
     val updateTime: Long,
-    val userId: Int
+    val userId: Long
 )
 
 data class Creator(
@@ -82,7 +82,7 @@ data class Creator(
     val province: Int,
     val remarkName: Any,
     val signature: String,
-    val userId: Int,
+    val userId: Long,
     val userType: Int,
     val vipType: Int
 )
